+++ mempool/txpool/legacypool/legacypool_test.go
@@ -543,7 +543,7 @@ func TestChainFork(t *testing.T) {
 	if _, err := pool.add(tx); err != nil {
 		t.Error("didn't expect error", err)
 	}
-	pool.RemoveTx(tx.Hash(), true, true)
+	pool.removeTx(tx.Hash(), true, true)
 
 	// reset the pool's internal state
 	resetState()
@@ -2589,6 +2589,61 @@ func TestSetCodeTransactionsReorg(t *testing.T) {
 	if err := pool.addRemoteSync(pricedTransaction(3, 100000, big.NewInt(1000), keyA)); err != nil {
 		t.Fatalf("failed to added single transaction: %v", err)
 	}
+}
+
+// TestRemoveTxTruncatePoolRace is a regression test for a race condition
+// between removing txs and runReorg loop. Run this with the -race flag to
+// ensure that there is no race condition between the two functions.
+func TestRemoveTxTruncatePoolRace(t *testing.T) {
+	statedb, _ := state.New(types.EmptyRootHash, state.NewDatabaseForTesting())
+	blockchain := newTestBlockChain(params.MergedTestChainConfig, 1000000, statedb, new(event.Feed))
+
+	pool := New(testTxPoolConfig, blockchain)
+	err := pool.Init(testTxPoolConfig.PriceLimit, blockchain.CurrentBlock(), newReserver())
+	if err != nil {
+		t.Fatalf("failed to init pool: %v", err)
+	}
+
+	// fill the pool with txs
+	fillPool(t, pool)
+
+	// make a copy of all hashes in the pool so that we do not have to iterate
+	// over pending and queue while we call RemoveTx, potentially triggering
+	// the race condition ourselves
+	var hashes []common.Hash
+	for _, txs := range pool.pending {
+		for _, tx := range txs.Flatten() {
+			hashes = append(hashes, tx.Hash())
+		}
+	}
+	for _, txs := range pool.queue {
+		for _, tx := range txs.Flatten() {
+			hashes = append(hashes, tx.Hash())
+		}
+	}
+
+	var wg sync.WaitGroup
+
+	// manually trigger the reorg loop to run (5 times just to ensure that we
+	// will trigger the race condition)
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		for range 5 {
+			pool.runReorg(make(chan struct{}), nil, nil, nil)
+		}
+	}()
+
+	// call RemoveTx on every tx in the pool
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		for _, hash := range hashes {
+			_ = pool.RemoveTx(hash, false, true)
+		}
+	}()
+
+	wg.Wait()
 }
 
 // Benchmarks the speed of validating the contents of the pending queue of the